replace mg.ExecRunFunc() with the pattern CmdCtx.WithCmd().Run()
The use-case is delegating to another command
and ExecRunFunc attempted to help by removing some of the boilerplate
but it's easy to mis-use when running a command that's being enhanced e.g. `go`.
ExecRunFunc would run the command OS command which works for that,
but the result isn't quite what's expected because when you run `go install`
from inside the editor, you expect linting to work and ExecRunFunc bypassed all that