Development (#65)
* add support for having multiple builtins with the same name

this is useful for cases where the editor calls a well-defined command that might be implemented by multiple reducers

* simplify and document mg.OutputStreams

* rename 'Completion' in golng.CompletionCtx and related types and funcs to Cursor* for clarity
rename golang.DedentCompletion to Dedent for clarity

* CursorCtx: add helpers FuncName() and MethodName() to return the func name under the cursor

* improve docs

* RunCmd: add helpers for more fun with flags

* replace mg.ExecRunFunc() with the pattern CmdCtx.WithCmd().Run()

The use-case is delegating to another command
and ExecRunFunc attempted to help by removing some of the boilerplate
but it's easy to mis-use when running a command that's being enhanced e.g. `go`.
ExecRunFunc would run the command OS command which works for that,
but the result isn't quite what's expected because when you run `go install`
from inside the editor, you expect linting to work and ExecRunFunc bypassed all that

* add new virtual command .actuate

this replaces the existing goto.definition and similar commands when
triggering a command using the mouse

* add new helpers mgutil.Clamp and mgutil.ClampPos

* replace buggy golang.clampSrcPos() with mgutil.ClampPos()

given src{} and pos==0, it'd return -1
given src{xy} and pos==3, it'd return 1/src{y} instead of 2/src{}

* add new helper mg.View.SrcPos()

* add new helper golang.NewViewCursorCtx()

* add support for the .actuate command to golang.Guru and golang.TestCmds

this helps https://github.com/DisposaBoy/GoSublime/issues/868
where both reducers are activated by the same user input (i.e. mouse click)