Fix a bug where a directory's vfs node would falsely claim that it's up-to-date.
This fixes a bug where type-checking would fail randomly due to a directory not having any .go files.

Files are left as-is as the side-effect is less severe.
Our *bluff* is called mainly due to creating the node after scanning a directory,
so the file mode is correct and we don't really care about precise mod times.