BuildContext.OpenFile: use ReadBlob to open files in the current package.
This should elimitate the constant re-reading of .go files from the disk when typing.
The use of PeekBlob helps, but doesn't cover the _test.go files unless we eventually load them, which we almost never do.

There is yet another (very) small increase in memory use, but it's limited to the package being edited.