vfs: Don't clear the root nodes when stat fails.
This appears to be the main cause of https://github.com/DisposaBoy/GoSublime/issues/957.
At least on Windows (maybe poor filepath->vfspath handling), we sometimes end up statting the root node,
and when it fails (because it has no path), we clear its nodes and thus remove all cached data.